Sweepify-Models1/requirements.txt
flask
torch
torchvision
pillow 
scikit-image 
requests
python-dotenv 
flask-cors
